Resolve #17 board.create_datetime, dev 설정 추가 (local mysql)

+++ cafe-core/src/main/java/net/kiel/cafe/repository/ArticleRepository.java
@@ -22,6 +22,6 @@ public interface ArticleRepository extends JpaRepository<Article, Long>, JpaSpec
     List<Article> findByBoard(Board board);
     Long countByBoard(Board board);
     
-    List<Article> findByBoardId(Integer boardId);
+    List<Article> findByBoardId(Long boardId);
     
 }
